Refactor error handling tests and remove unused test case
